Organise utils in a class
